Text Reader 2 now with text and bookmark retention
Now has some javascript to store the scrolled position and content of the text area, and also attempts to fit 80 characters on the full width of the available viewing area instead of 32px and cutting the control off at 80 when a checkbox is checked. 

Old version will remain where it is as it may be more useful if you don't want to mess with local storage and javascript.